requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
ifcopenshell>=0.7.0
plotly>=5.17.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.4.0
scikit-learn>=1.3.0

# Geometric computations
trimesh>=4.0.0
scipy>=1.11.0
shapely>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
